+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- Added [sleepy][] `-CountDown` option.'
+- Added [Add-HtmlHeader][] function.
+- Added [Resize-Window][] function.
+- Added [sleepy][] `-CountDown` option.
 - Added [Edit-Property][] Multiple EXPRESSIONS can be specified.
 - Added [Get-FullPath][] function.
+- Added [Get-RelativePath][] function.
 - Added [Invoke-Vivliostyle][] function.
 - Added [Get-Gmail][], [Get-Gcalendar][] -Detail option.
 - Added [sleepy][] -Until option.
@@ -394,6 +397,10 @@ Breaking Changes
 
 [Invoke-Vivliostyle]: src/Invoke-Vivliostyle_function.ps1
 [Get-FullPath]: src/Get-FullPath_function.ps1
+[Get-RelativePath]: src/Get-RelativePath_function.ps1
+[Resize-Window]: src/Resize-Window_function.ps1
+[Add-HtmlHeader]: src/Add-HtmlHeader_function.ps1
+
 
 [unreleased]: https://github.com/btklab/pwsh-sketches/compare/0.12.0..HEAD
 [0.12.0]: https://github.com/btklab/pwsh-sketches/releases/tag/0.12.0